Use `cubing.js` for case display.
https://cubing.net/api/visualcube/ has moved. It should continue to
work, but it will always be slower than local rendering. I'm trying to
encourage use of `cubing.js` as a replacement where possible, and it can
cover the use cases well (in addition to supporting many more features,
like animation and a wide variety of puzzles). This PR includes the
minimal changes to switch from VisualCube to `cubing.js`.

You may want additional options for the `<twisty-player>`, such as
`hint-facelets="none"` (which hides the hint facelets in the 3D view).